divide into 3 libraries
