+++ scripts/generate/templates/Slot.opts.js
