+++ scripts/generate/templates/StorageSlot.js
@@ -1,8 +1,8 @@
 const format = require('../format-lines');
-const { TYPES } = require('./StorageSlot.opts');
+const { TYPES } = require('./Slot.opts');
 
 const header = `\
-pragma solidity ^0.8.24;
+pragma solidity ^0.8.20;
 
 /**
  * @dev Library for reading and writing primitive types to specific storage slots.
@@ -28,87 +28,6 @@ pragma solidity ^0.8.24;
  * }
  * \`\`\`
  */
-`;
-
-const tooling = `\
-/**
- * @dev Derive an ERC-1967 slot from a string (namespace).
- */
-function erc1967slot(string memory namespace) internal pure returns (bytes32 slot) {
-  /// @solidity memory-safe-assembly
-  assembly {
-      slot := sub(keccak256(add(namespace, 0x20), mload(namespace)), 1)
-  }
-}
-
-/**
- * @dev Derive an ERC-7201 slot from a string (namespace).
- */
-function erc7201slot(string memory namespace) internal pure returns (bytes32 slot) {
-  /// @solidity memory-safe-assembly
-  assembly {
-      mstore(0x00, sub(keccak256(add(namespace, 0x20), mload(namespace)), 1))
-      slot := and(keccak256(0x00, 0x20), not(0xff))
-  }
-}
-
-/**
- * @dev Add an offset to a slot to get the n-th element of a structure or an array.
- */
-function offset(bytes32 slot, uint256 pos) internal pure returns (bytes32 result) {
-  unchecked {
-      return bytes32(uint256(slot) + pos);
-  }
-}
-
-/**
- * @dev Derive the location of the first element in an array from the slot where the length is stored.
- *
- * See https://docs.soliditylang.org/en/v0.8.20/internals/layout_in_storage.html#mappings-and-dynamic-arrays[Solidity docs for mappings and dynamic arrays.].
- */
-function deriveArray(bytes32 slot) internal pure returns (bytes32 result) {
-  /// @solidity memory-safe-assembly
-  assembly {
-      mstore(0x00, slot)
-      result := keccak256(0x00, 0x20)
-  }
-}
-`;
-
-const derive = ({ type }) => `\
-/**
- * @dev Derive the location of a mapping element from the key.
- *
- * See https://docs.soliditylang.org/en/v0.8.20/internals/layout_in_storage.html#mappings-and-dynamic-arrays[Solidity docs for mappings and dynamic arrays.].
- */
-function deriveMapping(bytes32 slot, ${type} key) internal pure returns (bytes32 result) {
-  /// @solidity memory-safe-assembly
-  assembly {
-      mstore(0x00, key)
-      mstore(0x20, slot)
-      result := keccak256(0x00, 0x40)
-  }
-}
-`;
-
-const derive2 = ({ type }) => `\
-/**
- * @dev Derive the location of a mapping element from the key.
- *
- * See https://docs.soliditylang.org/en/v0.8.20/internals/layout_in_storage.html#mappings-and-dynamic-arrays[Solidity docs for mappings and dynamic arrays.].
- */
-function deriveMapping(bytes32 slot, ${type} memory key) internal pure returns (bytes32 result) {
-  /// @solidity memory-safe-assembly
-  assembly {
-    let length := mload(key)
-    let begin :=  add(key, 0x20)
-    let end := add(begin, length)
-    let cache := mload(end)
-    mstore(end, slot)
-    result := keccak256(begin, add(length, 0x20))
-    mstore(end, cache)
-  }
-}
 `;
 
 const struct = ({ type, name }) => `\
@@ -139,70 +58,10 @@ function get${name}Slot(${type} storage store) internal pure returns (${name}Slo
 }
 `;
 
-const udvt = ({ type, name }) => `\
-/**
- * @dev UDVT that represent a slot holding a ${type}.
- */
-type ${name}SlotType is bytes32;
-
-/**
- * @dev Cast an arbitrary slot to a ${name}SlotType.
- */
-function as${name}Slot(bytes32 slot) internal pure returns (${name}SlotType) {
-  return ${name}SlotType.wrap(slot);
-}
-
-/**
- * @dev Load the value held at location \`slot\` in (normal) storage.
- */
-function sload(${name}SlotType slot) internal view returns (${type} value) {
-  /// @solidity memory-safe-assembly
-  assembly {
-    value := sload(slot)
-  }
-}
-
-/**
- * @dev Store \`value\` at location \`slot\` in (normal) storage.
- */
-function sstore(${name}SlotType slot, ${type} value) internal {
-  /// @solidity memory-safe-assembly
-  assembly {
-    sstore(slot, value)
-  }
-}
-
-/**
- * @dev Load the value held at location \`slot\` in transient storage.
- */
-function tload(${name}SlotType slot) internal view returns (${type} value) {
-  /// @solidity memory-safe-assembly
-  assembly {
-    value := tload(slot)
-  }
-}
-
-/**
- * @dev Store \`value\` at location \`slot\` in transient storage.
- */
-function tstore(${name}SlotType slot, ${type} value) internal {
-  /// @solidity memory-safe-assembly
-  assembly {
-    tstore(slot, value)
-  }
-}
-`;
-
 // GENERATE
 module.exports = format(
   header.trimEnd(),
   'library StorageSlot {',
-  '/// Derivation tooling',
-  tooling,
-  TYPES.flatMap(type => (type.isValueType ? derive(type) : derive2(type))),
-  '/// Storage slots as structs',
   TYPES.flatMap(type => [struct(type), type.isValueType ? '' : getStorage(type)]),
-  '/// Storage slots as udvt',
-  TYPES.filter(type => type.isValueType).map(type => udvt(type)),
   '}',
 );
