+++ test/utils/SlotDerivation.t.sol
@@ -1,139 +1,27 @@
 // SPDX-License-Identifier: MIT
-// This file was procedurally generated from scripts/generate/templates/StorageSlot.t.js.
+// This file was procedurally generated from scripts/generate/templates/SlotDerivation.t.js.
 
-pragma solidity ^0.8.24;
+pragma solidity ^0.8.20;
 
 import {Test} from "forge-std/Test.sol";
 
-import {StorageSlot} from "@openzeppelin/contracts/utils/StorageSlot.sol";
+import {SlotDerivation} from "@openzeppelin/contracts/utils/SlotDerivation.sol";
 
 // solhint-disable func-name-mixedcase
-contract StorageSlotTest is Test {
-    using StorageSlot for *;
-
-    address private _addressVariable;
-
-    function testValueAddress_1(address value) public {
-        bytes32 slot;
-        assembly {
-            slot := _addressVariable.slot
-        }
-
-        // set in solidity
-        _addressVariable = value;
-
-        // read using Slots
-        assertEq(slot.asAddressSlot().sload(), value);
-    }
-
-    function testValueAddress_2(address value) public {
-        bytes32 slot;
-        assembly {
-            slot := _addressVariable.slot
-        }
-
-        // set using Slots
-        slot.asAddressSlot().sstore(value);
-
-        // read in solidity
-        assertEq(_addressVariable, value);
-    }
-
-    bytes32 private _bytes32Variable;
-
-    function testValueBytes32_1(bytes32 value) public {
-        bytes32 slot;
-        assembly {
-            slot := _bytes32Variable.slot
-        }
-
-        // set in solidity
-        _bytes32Variable = value;
-
-        // read using Slots
-        assertEq(slot.asBytes32Slot().sload(), value);
-    }
-
-    function testValueBytes32_2(bytes32 value) public {
-        bytes32 slot;
-        assembly {
-            slot := _bytes32Variable.slot
-        }
-
-        // set using Slots
-        slot.asBytes32Slot().sstore(value);
-
-        // read in solidity
-        assertEq(_bytes32Variable, value);
-    }
-
-    uint256 private _uint256Variable;
-
-    function testValueUint256_1(uint256 value) public {
-        bytes32 slot;
-        assembly {
-            slot := _uint256Variable.slot
-        }
-
-        // set in solidity
-        _uint256Variable = value;
-
-        // read using Slots
-        assertEq(slot.asUint256Slot().sload(), value);
-    }
-
-    function testValueUint256_2(uint256 value) public {
-        bytes32 slot;
-        assembly {
-            slot := _uint256Variable.slot
-        }
-
-        // set using Slots
-        slot.asUint256Slot().sstore(value);
-
-        // read in solidity
-        assertEq(_uint256Variable, value);
-    }
-
-    int256 private _int256Variable;
-
-    function testValueInt256_1(int256 value) public {
-        bytes32 slot;
-        assembly {
-            slot := _int256Variable.slot
-        }
-
-        // set in solidity
-        _int256Variable = value;
-
-        // read using Slots
-        assertEq(slot.asInt256Slot().sload(), value);
-    }
-
-    function testValueInt256_2(int256 value) public {
-        bytes32 slot;
-        assembly {
-            slot := _int256Variable.slot
-        }
-
-        // set using Slots
-        slot.asInt256Slot().sstore(value);
-
-        // read in solidity
-        assertEq(_int256Variable, value);
-    }
+contract SlotDerivationTest is Test {
+    using SlotDerivation for bytes32;
 
     bytes[] private _array;
 
-    function testArray(uint256 length, uint256 offset) public {
+    function testDeriveArray(uint256 length, uint256 offset) public {
         length = bound(length, 1, type(uint256).max);
         offset = bound(offset, 0, length - 1);
 
         bytes32 baseSlot;
         assembly {
             baseSlot := _array.slot
+            sstore(baseSlot, length) // store length so solidity access does not revert
         }
-        baseSlot.asUint256Slot().sstore(length);
 
         bytes storage derived = _array[offset];
         bytes32 derivedSlot;
@@ -141,13 +29,12 @@ contract StorageSlotTest is Test {
             derivedSlot := derived.slot
         }
 
-        assertEq(baseSlot.asUint256Slot().sload(), _array.length);
         assertEq(baseSlot.deriveArray().offset(offset), derivedSlot);
     }
 
     mapping(address => bytes) private _addressMapping;
 
-    function testMappingAddress(address key) public {
+    function testDeriveMappingAddress(address key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _addressMapping.slot
@@ -164,7 +51,7 @@ contract StorageSlotTest is Test {
 
     mapping(bool => bytes) private _boolMapping;
 
-    function testMappingBoolean(bool key) public {
+    function testDeriveMappingBoolean(bool key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _boolMapping.slot
@@ -181,7 +68,7 @@ contract StorageSlotTest is Test {
 
     mapping(bytes32 => bytes) private _bytes32Mapping;
 
-    function testMappingBytes32(bytes32 key) public {
+    function testDeriveMappingBytes32(bytes32 key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _bytes32Mapping.slot
@@ -198,7 +85,7 @@ contract StorageSlotTest is Test {
 
     mapping(bytes4 => bytes) private _bytes4Mapping;
 
-    function testMappingBytes4(bytes4 key) public {
+    function testDeriveMappingBytes4(bytes4 key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _bytes4Mapping.slot
@@ -215,7 +102,7 @@ contract StorageSlotTest is Test {
 
     mapping(uint256 => bytes) private _uint256Mapping;
 
-    function testMappingUint256(uint256 key) public {
+    function testDeriveMappingUint256(uint256 key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _uint256Mapping.slot
@@ -232,7 +119,7 @@ contract StorageSlotTest is Test {
 
     mapping(uint32 => bytes) private _uint32Mapping;
 
-    function testMappingUint32(uint32 key) public {
+    function testDeriveMappingUint32(uint32 key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _uint32Mapping.slot
@@ -249,7 +136,7 @@ contract StorageSlotTest is Test {
 
     mapping(int256 => bytes) private _int256Mapping;
 
-    function testMappingInt256(int256 key) public {
+    function testDeriveMappingInt256(int256 key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _int256Mapping.slot
@@ -266,7 +153,7 @@ contract StorageSlotTest is Test {
 
     mapping(int32 => bytes) private _int32Mapping;
 
-    function testMappingInt32(int32 key) public {
+    function testDeriveMappingInt32(int32 key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _int32Mapping.slot
@@ -283,7 +170,7 @@ contract StorageSlotTest is Test {
 
     mapping(string => bytes) private _stringMapping;
 
-    function testMappingString(string memory key) public {
+    function testDeriveMappingString(string memory key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _stringMapping.slot
@@ -300,7 +187,7 @@ contract StorageSlotTest is Test {
 
     mapping(bytes => bytes) private _bytesMapping;
 
-    function testMappingBytes(bytes memory key) public {
+    function testDeriveMappingBytes(bytes memory key) public {
         bytes32 baseSlot;
         assembly {
             baseSlot := _bytesMapping.slot
